pyproject.toml
[project]
name = "applibry-api"
version = "0.1.0"
description = ""
authors = [
    {name = "AFOLABI GMATT MATTHEW",email = "gmattworld@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "mangum (>=0.19.0,<0.20.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "openai (>=1.100.2,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "boto3 (>=1.40.13,<2.0.0)",
    "standard-imghdr (>=3.13.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "applibry_api", from = "applibry_api"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
